Update behavior of `cmgr playtest`
The `playtest` command will now check the value of the
CMGR_INTERFACE environment variable and bind the web server
to that address if it is not '0.0.0.0' (in which case it will
bind only to 'localhost').  This behavior was chosen to keep
the common usecase (local testing) simple while still supporting
a means for exposing the instance to the network.