Update documentation
Adds additional documentation at both the top example directory
and for the `custom` example which explicitly documents constraints
and additions we make to the challenge's Dockerfile.  Specifically,
it documents `PUBLISH`, `LAUNCH`, and naming a stage `builder`.

Additionally, it changes "exposes" to "publishes" for the error
message presented when the challenge details/hints reference an
unknown server (no published ports) or unknown port name.