Disallow uppercase letters in 'namespace'
This fixes an issue identified by @syreal17 where uppercase letters
in the namespace are allowed initially but then break later actions.
Because the namespace is used as part of the Docker repository name,
we should not allow uppercase characters at all (Docker prohibits
them), so this removes them from the validation regex and updates
the error message.

Fixes #27