Implement FREEZE
Implements the `Freeze` API (#26) for allowing a defined stage (named
"base") to be pre-built and stored in a registry.  This registry is
then queried during build-time and used as the build cache if found.
The feature is currently only exposed in `cmgr` as it is intended to
be used during development rather than production (i.e., lock in a
known good state that production `cmgrd` servers will then pull from).