Force-remove images
Really the only case where this changes anything is if there happens to
be a stopped container referencing the image to be deleted. This
shouldn't usually happen when using cmgr though.

In that case, rather than raising an error, the stopped container's
image will simply be untagged (the required layers will continue to
exist and be referenced by SHA)