Provide capability for stable port assignments
This addresses the issue of an instances ports changing on restart
by providing a configuration knob to assign ports to `cmgr`.  If
the `CMGR_PORTS` environment variable is present, `cmgr` will
assume that it has the ability to fully control/assign the ports
in this range without further coordination with the OS or another
process.  This allows it to assign a port number to instances before
creating the instance with Docker which then enables Docker to keep
the same port between restarts of the same container.

Supersedes #31
Fixes #32