Refactor Dockerfile handling
Refactors the Dockerfile handling to decouple it some from the
base library.  Additionally, it adds the ability to export the
Dockerfiles embedded in the library as well as convert "typed"
challenges to a custom format by dropping in the Dockerfile it
currently uses into the challenge directory and then swapping
the challenge type to "custom".

Fixes #24